add log

+++ cmd/integration/commands/reset_state.go
@@ -140,10 +140,10 @@ func printStages(tx kv.Tx, snapshots *freezeblocks.RoSnapshots, borSn *freezeblo
 	fmt.Fprintf(w, "blocks.v2: %t, segments=%d, indices=%d\n", snapshots.Cfg().Enabled, snapshots.SegmentsMax(), snapshots.IndicesMax())
 	fmt.Fprintf(w, "blocks.bor.v2: segments=%d, indices=%d\n\n", borSn.SegmentsMax(), borSn.IndicesMax())
 
-	for i := 11499999; i < 15000002; i++ {
-		minV, _ := rawdbv3.TxNums.Min(tx, uint64(i))
-		maxV, _ := rawdbv3.TxNums.Max(tx, uint64(i))
-		fmt.Fprintf(w, "blockNumber=%d: min=%d, max=%d\n\n", uint64(i), minV, maxV)
+	for i := uint64(11499999); i < uint64(15000002); i++ {
+		minV, _ := rawdbv3.TxNums.Min(tx, i)
+		maxV, _ := rawdbv3.TxNums.Max(tx, i)
+		fmt.Fprintf(w, "blockNumber=%d: min=%d, max=%d\n\n", i, minV, maxV)
 	}
 
 	_, lastBlockInHistSnap, _ := rawdbv3.TxNums.FindBlockNum(tx, agg.EndTxNumMinimax())
